Improved back validations

+++ api/src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID')
 
 class ContactController {
 	async index(request, response) {
@@ -10,6 +11,10 @@ class ContactController {
 
 	async show(request, response) {
 		const { id } = request.params;
+
+    if(!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' });
+    }
 
 		const contact = await ContactsRepository.findById(id);
 
@@ -27,16 +32,22 @@ class ContactController {
 			return response.status(400).json({ error: 'Name is required' });
 		}
 
-		const contactExists = await ContactsRepository.findByEmail(email);
-		if (contactExists) {
-			return response.status(400).json({ error: 'This email is already in use' });
+    if (category_id && !isValidUUID(id)) {
+			return response.status(400).json({ error: 'Invalid category' });
 		}
+
+		if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists) {
+        return response.status(400).json({ error: 'This email is already in use' });
+		}
+    }
 
 		const contact = await ContactsRepository.create({
 			name,
-			email,
+			email: email || null,
 			phone,
-			category_id,
+			category_id: category_id || null,
 		});
 
 		response.status(201).json(contact);
@@ -46,26 +57,35 @@ class ContactController {
 		const { id } = request.params;
 		const { name, email, phone, category_id } = request.body;
 
+    if (isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' });
+    }
+
+    if (category_id && !isValidUUID(id)) {
+			return response.status(400).json({ error: 'Invalid category' });
+		}
+    if (!name) {
+      return response.status(400).json({ error: 'Name is required' });
+    }
+
 		const contactExists = await ContactsRepository.findById(id);
 
 		if (!contactExists) {
 			return response.status(404).json({ error: 'Contact not found' });
 		}
 
-		if (!name) {
-			return response.status(400).json({ error: 'Name is required' });
+		if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return response.status(400).json({ error: 'This email is already in use' });
 		}
-
-		const contactByEmail = await ContactsRepository.findByEmail(email);
-		if (contactByEmail && contactByEmail.id !== id) {
-			return response.status(400).json({ error: 'This email is already in use' });
-		}
+    }
 
 		const contact = await ContactsRepository.update(id, {
 			name,
-			email,
+			email: email || null,
 			phone,
-			category_id,
+			category_id: category_id || null,
 		});
 
 		response.json(contact);
@@ -73,6 +93,10 @@ class ContactController {
 
 	async delete(request, response) {
 		const { id } = request.params;
+
+    if (isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact id' });
+    }
 
 		await ContactsRepository.delete(id);
